+++ client/lib/src/pages/restart/controllers/restart_controller.dart
@@ -1,0 +1,37 @@
+import 'package:dio/dio.dart';
+import 'package:logging/logging.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+import '../../../providers/api_provider.dart';
+import '../../../widgets/controllers/action_controller.dart';
+
+part 'restart_controller.g.dart';
+
+@riverpod
+class RestartController extends _$RestartController with ActionController {
+  // ignore: avoid_public_notifier_properties
+  @override
+  String get action => 'restart unit $unit';
+
+  // ignore: avoid_public_notifier_properties
+  @override
+  Logger get logger => Logger('RestartController');
+
+  @override
+  ActionState build(String unit) {
+    ref.watch(systemdStatusApiClientProvider);
+    return ActionState.hint;
+  }
+
+  @override
+  Future<void> execute(CancelToken cancelToken) async {
+    final api = ref.read(systemdStatusApiClientProvider);
+    await api.unitsRestart(
+      unit,
+      $options: Options(
+        receiveTimeout: const Duration(minutes: 1, seconds: 30),
+      ),
+      $cancelToken: cancelToken,
+    );
+  }
+}