+++ client/lib/src/widgets/controllers/action_controller.dart
@@ -1,0 +1,34 @@
+import 'package:dio/dio.dart';
+import 'package:logging/logging.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+enum ActionState { hint, pending, success, failure }
+
+// ignore: invalid_use_of_internal_member
+mixin ActionController on BuildlessAutoDisposeNotifier<ActionState> {
+  @visibleForOverriding
+  String get action;
+
+  @visibleForOverriding
+  Logger get logger;
+
+  Future<void> execute(CancelToken cancelToken);
+
+  Future<void> call() async {
+    try {
+      state = ActionState.pending;
+      final cancelToken = CancelToken();
+      ref.onDispose(cancelToken.cancel);
+      await execute(cancelToken);
+      state = ActionState.success;
+    } on Exception catch (e, s) {
+      if (e case DioException(type: DioExceptionType.cancel)) {
+        state = ActionState.hint;
+        return;
+      }
+
+      logger.severe('Failed to $action', e, s);
+      state = ActionState.failure;
+    }
+  }
+}