+++ client/CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.0] - 2025-04-30
+### Added
+- Added reboot context menu action to reboot the server
+  - Requires special permission for the user that must be set on the server
+
 ### Fixed
 - Ensure autofill works for login page
-- Downgrade `file_picker` because of Android bug (See https://github.com/miguelpruivo/flutter_file_picker/issues/1769)
+- Downgrade `file_picker` because of Android bug (See https://github.com/miguelpruivo/flutter\_file\_picker/issues/1769)
 - Fix empty state of logs page
 
 ## [0.3.3] - 2025-04-29
@@ -110,7 +114,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/Skycoder42/systemd_status/compare/client%2Fv0.3.3...vHEAD
+[0.4.0]: https://github.com/Skycoder42/systemd_status/compare/client%2Fv0.3.3...v0.4.0
 [0.3.3]: https://github.com/Skycoder42/systemd_status/compare/client%2Fv0.3.2...v0.3.3
 [0.3.2]: https://github.com/Skycoder42/systemd_status/compare/client%2Fv0.3.1...v0.3.2
 [0.3.1]: https://github.com/Skycoder42/systemd_status/compare/client%2Fv0.3.0...v0.3.1