+++ server/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2025-04-30
+### Added
+- Add system endpoint to allow rebooting
+  - requires explicit `canReboot: true` permission on the user
+
+### Changed
+- Updated embedded web client to 0.4.0
+
 ## [0.3.3] - 2025-04-29
 ### Changed
 - Updated dependencies
@@ -58,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
+[0.4.0]: https://github.com/Skycoder42/systemd_status/compare/server%2Fv0.3.3...v0.4.0
 [0.3.3]: https://github.com/Skycoder42/systemd_status/compare/server%2Fv0.3.2...v0.3.3
 [0.3.2]: https://github.com/Skycoder42/systemd_status/compare/server%2Fv0.3.1...v0.3.2
 [0.3.1]: https://github.com/Skycoder42/systemd_status/compare/server%2Fv0.3.0...v0.3.1